Addon Manager: Refactor to extract metadata checker
